refactor : 포커스 이동 로직을 hook으로 분리
